requirements.txt
Django==1.9.8
celery==3.1.20
djangorestframework==3.3.1
redis==2.10.5
requests==2.7.0
psycopg2==2.8.4
jsonfield==2.0.2
enum==0.4.7
mock==3.0.5